+++ packages/controllers/contracts.go
@@ -49,7 +49,7 @@ func (c *Controller) Contracts() (string, error) {
 		if val[`wallet_id`] == `NULL` {
 			stateSmartLaws[ind][`wallet`] = ``
 		} else {
-			stateSmartLaws[ind][`wallet`] = lib.AddressToString(uint64(utils.StrToInt64(val[`wallet_id`])))
+			stateSmartLaws[ind][`wallet`] = lib.AddressToString(utils.StrToInt64(val[`wallet_id`]))
 		}
 		if val[`active`] == `NULL` {
 			stateSmartLaws[ind][`active`] = ``
