+++ packages/controllers/modal_anonym.go
@@ -41,7 +41,7 @@ func (c *Controller) ModalAnonym() (string, error) {
 	timeNow := utils.Time()
 
 	MyWalletData, err := c.OneRow("SELECT host, address_vote as addressVote  FROM dlt_wallets WHERE wallet_id = ?", c.SessWalletId).String()
-	MyWalletData[`address`] = lib.AddressToString(uint64(c.SessWalletId))
+	MyWalletData[`address`] = lib.AddressToString(c.SessWalletId)
 	if err != nil {
 		return "", utils.ErrInfo(err)
 	}
